requirements.txt
--only-binary=:all:
numpy==2.2.4
pyserial==3.5
smbus==1.1.post2
